address some code review comments
